Re #721 : MinGW 4.7 build system is now up and running (except for networkhuman module)
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@4838 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 6ed52a0b152da16fae7ecaacc3656fa4a5d5921b
Former-commit-id: 18343442c5252316e92572df01ebb3a98a89e68e

+++ src/libs/tgf/windowsspec.cpp
@@ -729,7 +729,7 @@ windowsGetOSInfo(int* pnMajor, int* pnMinor, int* pnBits)
 
     // Include build number.
     char buf[80];
-    sprintf(buf, " (build %d)", osvi.dwBuildNumber);
+    sprintf(buf, " (build %ld)", osvi.dwBuildNumber);
     strcat(pszVerSionString, buf);
 
     if (osvi.dwMajorVersion >= 6)
@@ -829,7 +829,7 @@ windowsSetThreadAffinity(int nCPUId)
 		// Affinity on a specified CPU => compute its mask (1 bit in the "system" mask).
 		int nCPUIndex = -1;
 		int nBitIndex = 0;
-		while (nBitIndex < sizeof(nSystemMask)*8 && nCPUIndex < nCPUId)
+		while (nBitIndex < (int)sizeof(nSystemMask)*8 && nCPUIndex < nCPUId)
 		{
 			if (nSystemMask & 1)
 				nCPUIndex++;
